wip: socket-default: Retry opening socket on ephemeral port in case of conflict
Since we now open sockets for each address family independently (via
IPV6_V6ONLY) and without SO_REUSEADDR, the order in which they are
opened doesn't matter anymore.

However, it could happen that opening the port that was allocated as
ephemeral port for one family is already used by a different process for
the second family.  Retry with a different port should that happen.

wip: if the kernel allocates the same port again after closing the socket,
we might have to keep failed ones open, add them to an array and close them
all once we succeeded